L-839, L-840 Move data serializion logic to core
