+++ packages/node/src/node.ts
@@ -1,5 +1,4 @@
 import { Duplex, Writable } from "stream";
-import { serializeError } from "serialize-error";
 
 import fetch from "cross-fetch";
 import { encode } from "@msgpack/msgpack";
@@ -64,16 +63,6 @@ export class Node extends Base {
     context: TContext = {} as TContext,
     stackContextHint?: StackContextHint,
   ) {
-    // Wrap context in an object, if it's not already
-    if (typeof context !== "object") {
-      const wrappedContext: unknown = { extra: context };
-      context = wrappedContext as TContext;
-    }
-    if (context instanceof Error) {
-      const wrappedContext: unknown = { error: context };
-      context = wrappedContext as TContext;
-    }
-
     // Process/sync the log, per `Base` logic
     context = { ...getStackContext(this, stackContextHint), ...context };
     const processedLog = await super.log(message, level, context);
@@ -98,93 +87,12 @@ export class Node extends Base {
   }
 
   private encodeAsMsgpack(logs: ILogtailLog[]): Buffer {
-    const maxDepth = this._options.contextObjectMaxDepth;
-    const logsWithISODateFormat = logs.map(log => ({
-      ...this.sanitizeForEncoding(log, maxDepth),
-      dt: log.dt.toISOString(),
-    }));
-    const encoded = encode(logsWithISODateFormat);
+    const encoded = encode(logs);
     const buffer = Buffer.from(
       encoded.buffer,
       encoded.byteOffset,
       encoded.byteLength,
     );
     return buffer;
-  }
-
-  private sanitizeForEncoding(
-    value: any,
-    maxDepth: number,
-    visitedObjects: WeakSet<any> = new WeakSet(),
-  ): any {
-    if (
-      value === null ||
-      typeof value === "boolean" ||
-      typeof value === "number" ||
-      typeof value === "string"
-    ) {
-      return value;
-    } else if (value instanceof Date) {
-      // Date instances can be invalid & toISOString() will fail
-      if (isNaN(value.getTime())) {
-        return value.toString();
-      }
-
-      return value.toISOString();
-    } else if (value instanceof Error) {
-      return serializeError(value);
-    } else if (
-      (typeof value === "object" || Array.isArray(value)) &&
-      (maxDepth < 1 || visitedObjects.has(value))
-    ) {
-      if (visitedObjects.has(value)) {
-        if (this._options.contextObjectCircularRefWarn) {
-          console.warn(
-            `[Logtail] Found a circular reference when serializing logs. Please do not use circular references in your logs.`,
-          );
-        }
-        return "<omitted circular reference>";
-      }
-      if (this._options.contextObjectMaxDepthWarn) {
-        console.warn(
-          `[Logtail] Max depth of ${this._options.contextObjectMaxDepth} reached when serializing logs. Please do not use excessive object depth in your logs.`,
-        );
-      }
-      return `<omitted context beyond configured max depth: ${this._options.contextObjectMaxDepth}>`;
-    } else if (Array.isArray(value)) {
-      visitedObjects.add(value);
-      const sanitizedArray = value.map(item =>
-        this.sanitizeForEncoding(item, maxDepth - 1, visitedObjects),
-      );
-      visitedObjects.delete(value);
-
-      return sanitizedArray;
-    } else if (typeof value === "object") {
-      const logClone: { [key: string]: any } = {};
-
-      visitedObjects.add(value);
-
-      Object.entries(value).forEach(item => {
-        const key = item[0];
-        const value = item[1];
-
-        const result = this.sanitizeForEncoding(
-          value,
-          maxDepth - 1,
-          visitedObjects,
-        );
-        if (result !== undefined) {
-          logClone[key] = result;
-        }
-      });
-
-      visitedObjects.delete(value);
-
-      return logClone;
-    } else if (typeof value === "undefined") {
-      return undefined;
-    } else {
-      return `<omitted unserializable ${typeof value}>`;
-    }
   }
 }
